install_templates: Add the default partition layout for the LS1021A-IOT
Set the "FDISK_PARTITION_LAYOUT" to use the special fdisk partition
layout file for the LS1021A-IOT target. Because this target need deploy the
u-boot into the SD card, and u-boot environment parameters are saved
in the 1MB. Need to reserve beginning 2MB space for the u-boot and
environment to avoid overriding the file system supperblock.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>