wr-core: Fix the host_package_install.sh script linux distribution mismatch issue
If user install the dpkg tool package in the RedHat system, or install
yum tool in the Debian system, then this script will find an error
distribution iformation for the "distro". Such as the RH system may be
recognized as the "Debian" system, and it will use the "apt-get" tool
to install the required packages.
This patch adds the distribution release limitation to make sure that
it could use the suitable package managers for the different distribution.

Signed-off-by: Guojian Zhou <guojian.zhou@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>