local.conf.sample: add ncurses-native ASSUME_PROVIDED
On Ubuntu 16.04 libsdl-dev depends on libcaca-dev which depends on
libcaca0 which depends on libncursesw5. Since we ASSUME_PROVIDED
libsdl-native this means when qemu-native runs configure and checks
for SDL we get a failure. The failure is caused since the linker will
attempt to use the host's libsdl which will use the host's libcaca but
_our_ libncurses. Since the symbol name mangling is different between
what the host's libcaca is looking for and what our libncurses is
providing the configure check for SDL will fail to compile. Once some
verbosity is enabled we see, the following type of errors:

//usr/lib/x86_64-linux-gnu/libcaca.so.0: undefined reference to `resize_term@NCURSESW_5.3.20021019'
//usr/lib/x86_64-linux-gnu/libcaca.so.0: undefined reference to `COLORS@NCURSESW_5.1.20000708'
//usr/lib/x86_64-linux-gnu/libcaca.so.0: undefined reference to `raw@NCURSES_TINFO_5.0.19991023'

On Ubuntu systems (or distro which have similar dependency chains)
this cross linking between host and our libraries can be avoided by
using the system's ncurses, thus by the addition being done here. For
non-Ubuntu systems which don't have this dependency chain and thus
don't exhibit this behavior this addition doesn't change anything.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>