apollolake, helios: change the submachine name from leafhill to apollolake
The machine name "leafhill" was already used by the scc in
kernel-meta/bsp/leafhill/leafhill-standard.scc. So set the KMACHINE
to that will cause that scc file to be used instead of the one
in our specific intel-apollolake layer. So use the apollolake as the
submachine name. It is also more suitable than the "leafhill" since
that is just one board name.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>